# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Brand Service API",
    "description": "Microservice for managing brand information, media assets, organization data, and AI-powered content analysis.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3005"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-to-service API key"
      }
    }
  },
  "paths": {
...
    "/internal/media-assets": {
      "get": {
        "summary": "Get all media assets for an organization",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "external_organization_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Media assets list"
          },
          "400": {
            "description": "Missing external_organization_id"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/openapi.json": {
      "get": {
        "summary": "OpenAPI specification",
        "responses": {
          "200": {
            "description": "OpenAPI JSON spec"
          },
          "404": {
            "description": "Spec not generated"
          }
        }
      }
    },
...
  }
}